requirements.txt
absl-py==1.4.0
abstra==0.33.1
abstra-runtimes==0.19.1
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
argilla==1.8.0
astroid==2.13.5
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
auto-gptq==0.2.2
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.39.1
black==22.12.0
blinker==1.6.2
boto3==1.26.155
botocore==1.29.155
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.22
click==8.1.3
clickhouse-connect==0.5.25
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colour==0.1.5
commonmark==0.9.1
compressed-rtf==1.0.6
contourpy==1.0.7
crontab==1.0.1
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.8
datasets==2.13.1
dearpygui==1.9.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.6
distlib==0.3.6
distro==1.8.0
docopt==0.6.2
duckdb==0.8.0
easygui==0.98.3
ebcdic==1.1.1
effdet==0.4.1
et-xmlfile==1.1.0
executing==1.2.0
extract-msg==0.41.1
faiss-cpu==1.7.4
fastapi==0.96.0
filelock==3.12.0
fire==0.5.0
Flask==2.3.2
Flask-Cors==4.0.0
flask-sock==0.6.0
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpt4all==0.2.3
greenlet==2.0.2
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
hnswlib==0.7.0
html2text==2020.1.16
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.15.1
humanfriendly==10.0
identify==2.5.24
idna==3.4
imageio==2.31.0
IMAPClient==2.3.1
iniconfig==2.0.0
InstructorEmbedding==1.0.1
iopath==0.1.10
ipython==8.10.0
isort==5.12.0
itsdangerous==2.1.2
jax==0.4.11
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
lambda==0.0.1
langchain==0.0.191
langchainplus-sdk==0.0.16
lark-parser==0.12.0
layoutparser==0.3.4
lazy-object-proxy==1.9.0
lazy_loader==0.2
libclang==16.0.0
llama-cpp-python==0.1.48
llama-index==0.6.1
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
ml-dtypes==0.2.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
msoffcrypto-tool==5.0.1
multidict==6.0.4
multipart==0.2.4
multiprocess==0.70.14
mypy==0.991
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.25.0
oauthlib==3.2.2
olefile==0.46
oletools==0.60.1
omegaconf==2.3.0
onnxruntime==1.15.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
pandoc==2.3
parso==0.8.3
pathspec==0.11.1
pcodedmp==1.2.6
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
pdftotree==0.5.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.5.1
pluggy==1.0.0
plumbum==1.8.2
ply==3.11
portalocker==2.7.0
posthog==3.0.1
pre-commit==3.2.0
progress==1.6
prompt-toolkit==3.0.38
protobuf==3.20.0
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.7.0
pylint==2.15.10
pypandoc==1.11
pyparsing==2.4.7
pypdf==3.9.1
PyPDF2==3.0.1
pyreadline3==3.4.1
PySimpleGUI==4.60.5
pytesseract==0.3.10
pytest==7.2.1
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
rake-nltk==1.0.6
red-black-tree-mod==1.20
regex==2023.6.3
requests==2.29.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.1
rouge==1.0.1
rsa==4.9
RTFDE==0.0.2
ruff==0.0.259
s3transfer==0.6.1
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
selectivesearch==0.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.1
simple-websocket==0.10.1
SimpleDocumentStore==0.1
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabula-py==2.7.0
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tiktoken==0.4.0
timm==0.9.2
tk==0.1.0
tokenizers==0.13.3
tomlkit==0.11.8
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
typer==0.9.0
types-docutils==0.20.0.1
types-pyOpenSSL==23.2.0.0
types-redis==4.5.5.0
types-requests==2.28.11.8
types-setuptools==67.1.0.0
types-urllib3==1.26.25.13
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.2
unstructured==0.6.6
unstructured-inference==0.5.1
urllib3==1.26.6
uvicorn==0.22.0
validators==0.20.0
virtualenv==20.23.0
Wand==0.6.11
watchfiles==0.19.0
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.5
win-unicode-console==0.5
wrapt==1.14.1
wsproto==1.2.0
XlsxWriter==3.1.2
xxhash==3.2.0
yarg==0.1.9
yarl==1.9.2
zstandard==0.21.0
